# docs/openapi.yaml
openapi: 3.0.0
info:
  title: ClimateCanary API
  version: 1.0.0
  description: API for out climate monitoring project with RBAC
...
paths:
...
  /rooms/{id}/violations:
    get:
      tags: [Climate Data]
      summary: Get historical violation log for a room
      description: "Returns the table of limit breaches seen below the charts in the UI.
        Needs CAN_VIEW_OWN_OFFICE_CLIMATE permission."
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: List of past warnings/violations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ViolationLogDTO' }
...
  /warnings:
    get:
      tags: [Climate Data]
      summary: Get all active climate warnings
      description: "Used for the global notification bell or dashboard alert list."
      responses:
        '200':
          description: List of current active warnings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/WarningCreateDTO' }
    post:
      tags: [Climate Data]
      summary: Raspberry Pi reports a limit violation
      description: "Hardware-facing endpoint. Triggered instantly when a sensor crosses a threshold."
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WarningCreateDTO' }
      responses:
        '201':
          description: Warning recorded and notification triggered

  # TIPS MANAGEMENT
...
  /buildings:
    get:
      tags: [Building Structure]
      summary: Get all buildings
      description: CAN_MANAGE_BUILDING_STRUCTURE permission required.
      responses:
        '200':
          description: List of buildings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BuildingDTO' }
    post:
      tags: [Building Structure]
      summary: Add a new building
      description: CAN_MANAGE_BUILDING_STRUCTURE permission required.
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BuildingCreateDTO' }
      responses:
        '201':
          description: Building created
  /buildings/{id}:
    patch:
      tags: [Building Structure]
      summary: Update building details
      description: CAN_MANAGE_BUILDING_STRUCTURE permission required.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BuildingCreateDTO' }
      responses:
        '200':
          description: Building updated
    delete:
      tags: [Building Structure]
      summary: Delete a building
      description: CAN_MANAGE_BUILDING_STRUCTURE permission required.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Building deleted
  /departments:
    get:
      tags: [Building Structure]
      summary: Get all departments for the table view
      description: CAN_MANAGE_BUILDING_STRUCTURE permission required.
      responses:
        '200':
          description: Returns departments with building names and room lists
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DepartmentDTO' }
    post:
      tags: [Building Structure]
      summary: Create a new department
      description: CAN_MANAGE_BUILDING_STRUCTURE permission required.
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DepartmentCreateDTO' }
      responses:
        '201':
          description: Department created
  /departments/{id}:
    patch:
      tags: [Building Structure]
      summary: Update department details
      description: CAN_MANAGE_BUILDING_STRUCTURE permission required.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DepartmentCreateDTO' }
      responses:
        '200':
          description: Department updated
    delete:
      tags: [Building Structure]
      summary: Remove a department 
      description: CAN_MANAGE_BUILDING_STRUCTURE permission required.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Department deleted
...
  /users:
    get:
      tags: [Users]
      summary: Get all users
      description: Requires CAN_MANAGE_USERS permission
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
        '401':
          description: Unauthorized

    post:
      tags: [Users]
      summary: Create new user
      description: Requires CAN_MANAGE_USERS permission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateDTO'
      responses:
        '201':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      description: Requires CAN_MANAGE_USERS permission
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Requested user 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'

    patch:
      tags: [Users]
      summary: Update user
      description: Requires CAN_MANAGE_USERS permission
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateDTO'
      responses:
        '200':
          description: Updated user

    delete:
      tags: [Users]
      summary: Delete user
      description: Requires CAN_MANAGE_USERS permission
      parameters:
         - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: User deleted

  # ROLE MANAGEMENT
  /roles:
    get:
      tags: [Roles]
      summary: Get all roles
      description: Requires CAN_MANAGE_USERS permission
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleDTO'

    patch:
      tags: [Roles]
      summary: Update role
      description: Requires CAN_MANAGE_USERS permission
      parameters:
        - $ref: '#/components/parameters/RoleId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdateDTO'
      responses:
        '200':
          description: Role updated

  # ABSENCES MANAGEMENT
...
components:
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RoleId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    PermissionEnum:
      type: string
      description: "All possible permission constants"
      enum:
        - CAN_VIEW_OWN_OFFICE_CLIMATE
        - CAN_VIEW_OWN_SHARED_CLIMATE
        - CAN_MANAGE_OWN_ABSENCE
        - CAN_VIEW_OWN_DEPARTMENT_MEASURES
        - CAN_VIEW_OWN_DEPARTMENT_LIMITS_VIOLATION
        - CAN_VIEW_ABSENCE_VIEW
        - CAN_VIEW_COMPANY_AGGR
        - CAN_VIEW_VIOLATIONS_PER_DEPARTMENT
        - CAN_CHANGE_LIMITS
        - CAN_MANAGE_TIPS
        - CAN_VIEW_ALL_ROOMS
        - CAN_MANAGE_USERS
        - CAN_MANAGE_BUILDING_STRUCTURE
        - CAN_MANAGE_DEVICES
    UserDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employeeId:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleDTO'

    # what Frontend sends for creating of an entity
    UserCreateDTO:
      type: object
      required: [username, password, employeeId, secondName]
      properties:
        employeeId:
          type: string
          description: "The unique company employee number"
        username:
          type: string
        password:
          type: string
          format: password
          description: "Plain text password (will be hashed with bcrypt by the server)"
        firstName:
          type: string
        secondName:
          type: string
        roleIds:
          type: array
          items:
            type: string
            format: uuid
          description: "List of Role UUIDs to assign to the new user"
    RoleDTO:
      type: object
      required: [id, name]
      properties:
        id: 
          type: string
          format: uuid
          description: "Unique identifier from the Entity"
        name: 
          type: string
        permissions: 
          type: array
          description: "List of permission strings assigned to this role"
          items: 
            $ref: '#/components/schemas/PermissionEnum'
    RoleUpdateDTO:
      type: object
      description: "Fields for updating an existing role. ID is excluded as it is in the path."
      properties:
        name: 
          type: string
        permissions: 
          type: array
          description: "List of permission strings assigned to this role"
          items: 
            $ref: '#/components/schemas/PermissionEnum'
...
    BuildingDTO:
      type: object
      required: [id, name, address]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: "HQ" }
        address:
          type: string
          example: "Main Street 5, 6020 Innsbruck"
    DepartmentDTO:
      type: object
      required: [id, name, buildingId]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: "IT" }
        buildingId: { type: string, format: uuid }
        buildingName: { type: string, example: "HQ" } 
        roomNumbers: 
          type: array
          items: { type: string }
...
    BuildingCreateDTO:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Finance Centre West"
        address:
          type: string
          example: "Main Street 5, 6020 Innsbruck"
    DepartmentCreateDTO:
      type: object
      required: [name, buildingId]
      properties:
        name:
          type: string
          example: "HR"
        buildingId:
          type: string
          format: uuid
          description: "The UUID of the building this department belongs to"
...
    WarningCreateDTO:
      type: object
      required: [roomId, sensorType, value, limitExceeded]
      description: Sent by Pi ONLY when a limit is exceeded
      properties:
        roomId: { type: string, format: uuid }
        sensorType: { type: string, enum: [TEMPERATURE, HUMIDITY, CO2] }
        value: { type: number, example: 1200 }
        limitExceeded: { type: number, example: 1000 }
        message: { type: string, example: "Critical CO2 level in Room 101" }
...
    ViolationLogDTO:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        sensorType: { type: string, enum: [TEMPERATURE, HUMIDITY, CO2] }
        value: { type: number }
        limitViolated: { type: number }
        duration: { type: string, example: "15 min" }
...